[TM_WEB-10] Add global task context
