+++ react-dashboard/pages/_app.js
@@ -1,0 +1,9 @@
+import { TaskProvider } from '../context/TaskContext'
+
+export default function MyApp({ Component, pageProps }) {
+  return (
+    <TaskProvider>
+      <Component {...pageProps} />
+    </TaskProvider>
+  )
+}
